+++ magda-web-client/src/Components/Dataset/Add/DatasetAddCommon.ts
@@ -1,12 +1,12 @@
 import uuidv4 from "uuid/v4";
 
-import { Contact } from "Components/Editing/Editors/contactEditor";
 import {
     licenseLevel,
     ContactPointDisplayOption
 } from "constants/DatasetConstants";
 import { fetchOrganization } from "api-clients/RegistryApis";
 import { config } from "config";
+import { User } from "reducers/userManagementReducer";
 
 export type File = {
     title: string;
@@ -71,7 +71,8 @@ export type Dataset = {
     languages?: string[];
     keywords?: string[];
     themes?: string[];
-    contactPointFull?: Contact[];
+    owningOrgUnitId?: string;
+    contactPointDisplay?: string;
     publisher?: OrganisationAutocompleteChoice;
     landingPage?: string;
     importance?: string;
@@ -138,13 +139,14 @@ type Access = {
     downloadURL?: string;
 };
 
-function createBlankState(): State {
+function createBlankState(user: User): State {
     return {
         files: [],
         processing: false,
         dataset: {
             title: "Untitled",
-            languages: ["eng"]
+            languages: ["eng"],
+            owningOrgUnitId: user.orgUnitId
         },
         datasetPublishing: {
             state: "draft",
@@ -168,13 +170,13 @@ function createBlankState(): State {
 
 // saving data in the local storage for now
 // TODO: consider whether it makes sense to store this in registery as a custom state or something
-export async function loadState(id: string): Promise<State> {
+export async function loadState(id: string, user: User): Promise<State> {
     const stateString = localStorage[id];
     let state: State;
     if (stateString) {
         state = JSON.parse(stateString);
     } else {
-        state = createBlankState();
+        state = createBlankState(user);
     }
 
     if (