+++ magda-web-client/src/Components/Dataset/Add/DatasetAddMetadataPage.tsx
@@ -46,6 +46,7 @@ import datasetDistributionsAspect from "@magda/registry-aspects/dataset-distribu
 import dcatDistributionStringsAspect from "@magda/registry-aspects/dcat-distribution-strings.schema.json";
 import usageAspect from "@magda/registry-aspects/usage.schema.json";
 import accessAspect from "@magda/registry-aspects/access.schema.json";
+import datasetAccessControl from "@magda/registry-aspects/dataset-access-control.schema.json";
 
 import "./DatasetAddMetadataPage.scss";
 import "./DatasetAddFilesPage.scss";
@@ -71,7 +72,8 @@ const aspects = {
     "dataset-distributions": datasetDistributionsAspect,
     "dcat-distribution-strings": dcatDistributionStringsAspect,
     usage: usageAspect,
-    access: accessAspect
+    access: accessAspect,
+    "dataset-access-control": datasetAccessControl
 };
 
 type Props = {
@@ -436,24 +438,33 @@ class NewDataset extends React.Component<Props, State> {
         return (
             <div className="row dataset-access-and-use-page">
                 <div className="col-sm-12">
-                    <h2>Dataset access and use</h2>
-                    <hr />
-                    <h3>User access</h3>
-                    <h4>Who can see the dataset once it is published?</h4>
-                    <ToolTip>
-                        We recommend you publish your data to everyone in your
-                        organisation to help prevent data silos.
-                    </ToolTip>
-                    <p>
-                        <AlwaysEditor
-                            value={datasetPublishing.level}
-                            onChange={editDatasetPublishing("level")}
-                            editor={codelistRadioEditor(
-                                "dataset-publishing-level",
-                                codelists.publishingLevel
-                            )}
-                        />
-                    </p>
+                    <h2>Access and Use</h2>
+                    <h3 className="with-underline">User access</h3>
+                    <div className="question-who-can-see-dataset">
+                        <h4 className="with-icon">
+                            <span>
+                                Who can see the dataset once it is published?
+                            </span>
+                            <span className="help-icon-container">
+                                <img src={helpIcon} />
+                            </span>
+                        </h4>
+                        <ToolTip>
+                            We recommend you publish your data to everyone in
+                            your organisation to help prevent data silos.
+                        </ToolTip>
+                        <div>
+                            <AlwaysEditor
+                                value={datasetPublishing.level}
+                                onChange={editDatasetPublishing("level")}
+                                editor={codelistRadioEditor(
+                                    "dataset-publishing-level",
+                                    codelists.publishingLevel
+                                )}
+                            />
+                        </div>
+                    </div>
+
                     <h4>How can other users access this dataset?</h4>
                     <ToolTip>
                         Include locations on share drives, URLs of databases,
@@ -790,6 +801,9 @@ class NewDataset extends React.Component<Props, State> {
                 usage,
                 "dataset-publisher": {
                     publisher: publisherId
+                },
+                "dataset-access-control": {
+                    orgUnitOwnerId: dataset.owningOrgUnitId
                 }
             }
         };