+++ magda-web-client/src/actions/recordActions.ts
@@ -4,7 +4,6 @@ import { actionTypes } from "../constants/ActionTypes";
 import { RecordAction, RawDataset } from "../helpers/record";
 import { FetchError } from "../types";
 import request from "helpers/request";
-import datasetAccessControlAspect from "@magda/registry-aspects/dataset-access-control.schema.json";
 
 export function requestDataset(id: string): RecordAction {
     return {
@@ -240,28 +239,6 @@ export function createRecord(
         dispatch(createNewDataset(inputDataset));
         try {
             // -- set up access control aspect
-            const state = getState();
-            aspects["dataset-access-control"] = datasetAccessControlAspect;
-            if (!inputDataset["aspects"]["dataset-access-control"]) {
-                inputDataset["aspects"]["dataset-access-control"] = {};
-            }
-            if (
-                state.userManagement &&
-                state.userManagement.user &&
-                state.userManagement.user.id
-            ) {
-                inputDataset["aspects"]["dataset-access-control"]["ownerId"] =
-                    state.userManagement.user.id;
-            }
-            if (
-                state.userManagement &&
-                state.userManagement.user &&
-                state.userManagement.user.orgUnitId
-            ) {
-                inputDataset["aspects"]["dataset-access-control"][
-                    "orgUnitOwnerId"
-                ] = state.userManagement.user.orgUnitId;
-            }
             for (const [aspect, definition] of Object.entries(aspects)) {
                 await ensureAspectExists(aspect, definition);
             }