+++ magda-web-client/src/constants/DatasetConstants.ts
@@ -559,9 +559,10 @@ export const contactPointDisplay: Record<ContactPointDisplayOption, string> = {
 };
 
 export const publishingLevel = {
-    agency: "Everyone in my organisation",
+    agency: "Everyone in my organisation (recommended)",
+    branch: "My branch only",
+    section: "My section only",
     team: "My team only"
-    // public: "Everyone (Open Data)"
 };
 
 // TODO: this is a preconfigured agency level configured license lookup feature being faked here